refactor: transaction template 대신 @Transactional 적용
