+++ src/main/java/datastreams_knu/bigpicture/interest/agent/InterestCrawlingAgent.java
@@ -1,23 +1,19 @@
 package datastreams_knu.bigpicture.interest.agent;
 
-import datastreams_knu.bigpicture.common.config.AiModelConfig;
+import datastreams_knu.bigpicture.common.dto.CrawlingResultDto;
 import datastreams_knu.bigpicture.common.util.WebClientUtil;
 import datastreams_knu.bigpicture.interest.agent.dto.DateRangeDto;
-import datastreams_knu.bigpicture.interest.agent.dto.InterestCrawlingResultDto;
 import datastreams_knu.bigpicture.interest.agent.dto.KoreaInterestCrawlingDto;
 import datastreams_knu.bigpicture.interest.agent.dto.USInterestCrawlingDto;
-import datastreams_knu.bigpicture.interest.domain.KoreaInterest;
-import datastreams_knu.bigpicture.interest.domain.USInterest;
+import datastreams_knu.bigpicture.interest.entity.KoreaInterest;
+import datastreams_knu.bigpicture.interest.entity.USInterest;
 import datastreams_knu.bigpicture.interest.repository.KoreaInterestRepository;
 import datastreams_knu.bigpicture.interest.repository.USInterestRepository;
 import dev.langchain4j.agent.tool.Tool;
-import dev.langchain4j.model.chat.ChatLanguageModel;
-import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDate;
 import java.time.YearMonth;
@@ -34,9 +30,6 @@ public class InterestCrawlingAgent {
     private final WebClientUtil webClientUtil;
     private final KoreaInterestRepository koreaInterestRepository;
     private final USInterestRepository usInterestRepository;
-    private final PlatformTransactionManager transactionManager;
-
-    private TransactionTemplate txTemplate;
 
     @Value("${ecos.api.base-url}")
     private String ecosBaseUrl;
@@ -48,12 +41,6 @@ public class InterestCrawlingAgent {
     @Value("${fred.api.key}")
     private String fredApiKey;
 
-    @PostConstruct
-    private void init() {
-        this.txTemplate = new TransactionTemplate(transactionManager);
-    }
-
-
     @Tool("지난 n년 동안의 한국 금리를 크롤링합니다.")
     public KoreaInterestCrawlingDto crawlingInterestsOfKorea(int n) {
         String dateRange = getKoreaDateRange(n);
@@ -61,22 +48,21 @@ public class InterestCrawlingAgent {
         return webClientUtil.get(url, KoreaInterestCrawlingDto.class);
     }
 
+    @Transactional
     @Tool("크롤링 된 지난 n년 동안의 한국 금리를 DB에 저장하고 처리 성공 여부를 반환합니다.")
-    public InterestCrawlingResultDto saveKoreaInterest(KoreaInterestCrawlingDto koreaInterestCrawlingDto) {
-        return txTemplate.execute((status) -> {
-            try {
-                List<StatisticRow> interestDataRows = koreaInterestCrawlingDto.getStatisticSearch().getRow();
-                for (StatisticRow row : interestDataRows) {
-                    LocalDate interestDate = parseStringToLocalDate(row.getTime());
-                    Float interestRate = Float.parseFloat(row.getDataValue());
-                    KoreaInterest koreaInterest = KoreaInterest.of(interestDate, interestRate);
-                    koreaInterestRepository.save(koreaInterest);
-                }
-                return InterestCrawlingResultDto.of(true, "성공적으로 금리를 크롤링하였습니다.");
-            } catch (Exception e) {
-                return InterestCrawlingResultDto.of(false, "금리 크롤링을 실패하였습니다.");
+    public CrawlingResultDto saveKoreaInterest(KoreaInterestCrawlingDto koreaInterestCrawlingDto) {
+        try {
+            List<StatisticRow> interestDataRows = koreaInterestCrawlingDto.getStatisticSearch().getRow();
+            for (StatisticRow row : interestDataRows) {
+                LocalDate interestDate = parseStringToLocalDate(row.getTime());
+                Float interestRate = Float.parseFloat(row.getDataValue());
+                KoreaInterest koreaInterest = KoreaInterest.of(interestDate, interestRate);
+                koreaInterestRepository.save(koreaInterest);
             }
-        });
+            return CrawlingResultDto.of(true, "성공적으로 금리를 크롤링하였습니다.");
+        } catch (Exception e) {
+            return CrawlingResultDto.of(false, "금리 크롤링을 실패하였습니다.");
+        }
     }
 
     @Tool("지난 n년 동안의 미국 금리를 크롤링합니다.")
@@ -86,22 +72,21 @@ public class InterestCrawlingAgent {
         return webClientUtil.get(url, USInterestCrawlingDto.class);
     }
 
+    @Transactional
     @Tool("크롤링 된 지난 n년 동안의 미국 금리를 DB에 저장하고 처리 성공 여부를 반환합니다.")
-    public InterestCrawlingResultDto saveUSInterest(USInterestCrawlingDto usInterestCrawlingDto) {
-        return txTemplate.execute((status) -> {
-            try {
-                List<ObservationsRow> rows = usInterestCrawlingDto.getObservations();
-                for (ObservationsRow row : rows) {
-                    LocalDate interestDate = parseStringToLocalDate(row.getTime());
-                    Float interestRate = Float.parseFloat(row.getDataValue());
-                    USInterest usInterest = USInterest.of(interestDate, interestRate);
-                    usInterestRepository.save(usInterest);
-                }
-                return InterestCrawlingResultDto.of(true, "성공적으로 금리를 크롤링하였습니다.");
-            } catch (Exception e) {
-                return InterestCrawlingResultDto.of(false, "금리 크롤링을 실패하였습니다.");
+    public CrawlingResultDto saveUSInterest(USInterestCrawlingDto usInterestCrawlingDto) {
+        try {
+            List<ObservationsRow> rows = usInterestCrawlingDto.getObservations();
+            for (ObservationsRow row : rows) {
+                LocalDate interestDate = parseStringToLocalDate(row.getTime());
+                Float interestRate = Float.parseFloat(row.getDataValue());
+                USInterest usInterest = USInterest.of(interestDate, interestRate);
+                usInterestRepository.save(usInterest);
             }
-        });
+            return CrawlingResultDto.of(true, "성공적으로 금리를 크롤링하였습니다.");
+        } catch (Exception e) {
+            return CrawlingResultDto.of(false, "금리 크롤링을 실패하였습니다.");
+        }
     }
 
     private static LocalDate parseStringToLocalDate(String dateString) {
