+++ src/main/java/datastreams_knu/bigpicture/news/agent/NewsCrawlingAgent.java
@@ -3,11 +3,15 @@ package datastreams_knu.bigpicture.news.agent;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import datastreams_knu.bigpicture.common.config.AiModelConfig;
+import datastreams_knu.bigpicture.common.dto.CrawlingResultDto;
 import datastreams_knu.bigpicture.common.dto.DateRangeDto;
 import datastreams_knu.bigpicture.common.util.WebClientUtil;
-import datastreams_knu.bigpicture.news.agent.dto.*;
-import datastreams_knu.bigpicture.news.domain.News;
-import datastreams_knu.bigpicture.news.domain.NewsInfo;
+import datastreams_knu.bigpicture.news.agent.dto.CrawledNewsDto;
+import datastreams_knu.bigpicture.news.agent.dto.StringSummaryDto;
+import datastreams_knu.bigpicture.news.agent.dto.SummarizedMultipleNewsDto;
+import datastreams_knu.bigpicture.news.agent.dto.SummarizedNewsDto;
+import datastreams_knu.bigpicture.news.entity.News;
+import datastreams_knu.bigpicture.news.entity.NewsInfo;
 import datastreams_knu.bigpicture.news.exception.NewsCrawlingException;
 import datastreams_knu.bigpicture.news.repository.NewsRepository;
 import dev.langchain4j.agent.tool.Tool;
@@ -21,8 +25,7 @@ import org.jsoup.nodes.Document;
 import org.jsoup.nodes.Element;
 import org.jsoup.select.Elements;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
 import java.net.URLEncoder;
@@ -46,15 +49,12 @@ public class NewsCrawlingAgent {
     private final AiModelConfig aiModelConfig;
     private final ObjectMapper objectMapper;
     private final NewsRepository newsRepository;
-    private final PlatformTransactionManager transactionManager;
 
     private ChatLanguageModel model;
-    private TransactionTemplate txTemplate;
 
     @PostConstruct
     public void init() {
         this.model = aiModelConfig.openAiChatModel();
-        this.txTemplate = new TransactionTemplate(transactionManager);
     }
 
     @Tool("keyword를 기반으로 지난 7일 동안의 관련 뉴스 기사를 검색합니다.")
@@ -109,17 +109,16 @@ public class NewsCrawlingAgent {
         }
     }
 
+    @Transactional
     @Tool("생성된 기사 요약과 기사 출처 URL들 그리고 해당 기사의 날짜들을 DB에 저장하고 처리 성공 여부를 반환합니다.")
-    public NewsCrawlingResultDto saveSummarizedNewsWithUrls(String keyword, SummarizedMultipleNewsDto summarizedMultipleNews) throws Exception {
-        return txTemplate.execute((status) -> {
-            try {
-                News news = updateNews(keyword, summarizedMultipleNews);
-                newsRepository.save(news);
-                return NewsCrawlingResultDto.of(true, "성공적으로 기사를 크롤링하였습니다.");
-            } catch (Exception e) {
-                return NewsCrawlingResultDto.of(false, "기사 크롤링을 실패하였습니다.");
-            }
-        });
+    public CrawlingResultDto saveSummarizedNewsWithUrls(String keyword, SummarizedMultipleNewsDto summarizedMultipleNews) {
+        try {
+            News news = updateNews(keyword, summarizedMultipleNews);
+            newsRepository.save(news);
+            return CrawlingResultDto.of(true, "성공적으로 기사를 크롤링하였습니다.");
+        } catch (Exception e) {
+            return CrawlingResultDto.of(false, "기사 크롤링을 실패하였습니다.");
+        }
     }
 
     private static String createUrl(Map<String, String> params) {
