+++ src/main/java/datastreams_knu/bigpicture/stock/agent/StockCrawlingAgent.java
@@ -1,21 +1,19 @@
 package datastreams_knu.bigpicture.stock.agent;
 
+import datastreams_knu.bigpicture.common.dto.CrawlingResultDto;
 import datastreams_knu.bigpicture.common.dto.DateRangeDto;
 import datastreams_knu.bigpicture.common.util.WebClientUtil;
 import datastreams_knu.bigpicture.stock.agent.dto.KoreaStockCrawlingDto;
-import datastreams_knu.bigpicture.stock.agent.dto.StockCrawlingResultDto;
 import datastreams_knu.bigpicture.stock.agent.dto.StockInfoDto;
 import datastreams_knu.bigpicture.stock.agent.dto.USStockCrawlingDto;
 import datastreams_knu.bigpicture.stock.domain.Stock;
 import datastreams_knu.bigpicture.stock.domain.StockInfo;
 import datastreams_knu.bigpicture.stock.repository.StockRepository;
 import dev.langchain4j.agent.tool.Tool;
-import jakarta.annotation.PostConstruct;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import org.springframework.transaction.PlatformTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.net.URLEncoder;
 import java.nio.charset.StandardCharsets;
@@ -33,9 +31,6 @@ public class StockCrawlingAgent {
 
     private final WebClientUtil webClientUtil;
     private final StockRepository stockRepository;
-    private final PlatformTransactionManager transactionManager;
-
-    private TransactionTemplate txTemplate;
 
     @Value("${korea-stock.api.base-url}")
     private String koreaStockBaseUrl;
@@ -46,11 +41,6 @@ public class StockCrawlingAgent {
     private String usStockBaseUrl;
     @Value("${us-stock.api.key}")
     private String usStockApiKey;
-
-    @PostConstruct
-    public void init() {
-        this.txTemplate = new TransactionTemplate(transactionManager);
-    }
 
     @Tool("주어진 'stockName'에 해당하는 한국 주식의 주가 데이터를 수집합니다.")
     public List<StockInfoDto> crawlingKoreaStockByName(String stockName) {
@@ -87,27 +77,25 @@ public class StockCrawlingAgent {
             .collect(Collectors.toList());
     }
 
-
+    @Transactional
     @Tool("크롤링 된 주가 데이터를 DB에 저장하고 처리 성공 여부를 반환합니다.")
-    public StockCrawlingResultDto saveStock(String stockName, List<StockInfoDto> stockInfos) {
-        return txTemplate.execute((status) -> {
-            try {
-                Stock stock = stockRepository.findByName(stockName)
-                    .orElse(Stock.of(stockName));
+    public CrawlingResultDto saveStock(String stockName, List<StockInfoDto> stockInfos) {
+        try {
+            Stock stock = stockRepository.findByName(stockName)
+                .orElse(Stock.of(stockName));
 
-                stock.clearStockInfos();
+            stock.clearStockInfos();
 
-                stockInfos.stream()
-                    .map(info -> StockInfo.of(info.getStockPrice(), info.getStockDate()))
-                    .forEach(stock::addStockInfo);
+            stockInfos.stream()
+                .map(info -> StockInfo.of(info.getStockPrice(), info.getStockDate()))
+                .forEach(stock::addStockInfo);
 
-                stockRepository.save(stock);
+            stockRepository.save(stock);
 
-                return StockCrawlingResultDto.of(true, "성공적으로 주식을 크롤링하였습니다.");
-            } catch (Exception e) {
-                return StockCrawlingResultDto.of(false, "주식 크롤링을 실패하였습니다.");
-            }
-        });
+            return CrawlingResultDto.of(true, "성공적으로 주식을 크롤링하였습니다.");
+        } catch (Exception e) {
+            return CrawlingResultDto.of(false, "주식 크롤링을 실패하였습니다.");
+        }
     }
 
     private String createKoreaStockUrl(String stockName, DateRangeDto dateRange) {
