+++ src/main/java/datastreams_knu/bigpicture/stock/entity/Stock.java
