+++ src/main/java/datastreams_knu/bigpicture/stock/entity/StockInfo.java
@@ -9,7 +9,6 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
-import java.time.Instant;
 import java.time.LocalDate;
 
 @Getter
